requirements.txt
# ============================================================
# Core API & async
# ============================================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-multipart==0.0.12          # file uploads
python-dotenv==1.0.1
httpx==0.27.2
pyyaml==6.0.2

# ============================================================
# LLM providers
# ============================================================
openai==1.54.4
anthropic==0.39.0
groq==0.11.0
# Ollama uses plain HTTP — no extra package needed

# ============================================================
# Knowledge / Vector DB
# ============================================================
chromadb==0.5.18
pypdf==5.1.0                      # PDF ingestion
python-docx==1.1.2                # DOCX ingestion
openpyxl==3.1.5                   # XLSX read/write

# ============================================================
# Web scraping
# ============================================================
beautifulsoup4==4.12.3

# ============================================================
# Redis (distributed task queue)
# ============================================================
redis==5.2.0

# ============================================================
# Utilities
# ============================================================
structlog==24.4.0
